Add ListView snapMode to ensure flicking settles on an item boundary.

+++ src/declarative/graphicsitems/qmlgraphicsflickable_p_p.h
@@ -111,8 +111,8 @@ public:
     QTime velocityTime;
     QPointF lastFlickablePosition;
     qreal reportedVelocitySmoothing;
-    int flickTargetX;
-    int flickTargetY;
+    qreal flickTargetX;
+    qreal flickTargetY;
     QGraphicsSceneMouseEvent *delayedPressEvent;
     QGraphicsItem *delayedPressTarget;
     QBasicTimer delayedPressTimer;
